requirements.txt
django>=4.2,<5
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-cors-headers>=4.3
Pillow>=10.0
